Update 09.markmin
I removed the "scope=... " line which produced errors and seems redundant now. 
I added **args for first_name etc which seems to be how facebook does it now. (from stackoverflow.com/questions/16669977)
I tested the code on my server and it seems to work unlike the previous which gave errors.